Some corrections for provinces.yaml
Hello, these changes that I've submitted here are minor but they surely are made to more accurately present some of the cities.
Toyohara (meaning "bountiful plain") was the Japanese name for today's Yuzhno-Sakhalinsk (which literally means South Sakhalin City), the Japanese never called it Yujinosaharinsuku before losing the war cause that's just the japanisation of the name the city holds today.
The changes I made to the city of Cluj-Napoca are also historical. The "Napoca" part of the city's name was added in 1974 by the communists because they wished to further solidify the city in Romanian culture, Napoca being the first ever settlement ever build here by the Dacians. Being romanian myself, I couldn't help but be slightly bothered by the modern name placed upon the city in this timeframe. I believe those were the two most important changes I did and believed I needed to provide context behind. Thank you for making this mod as it's something I wished to see in HOI4 for a long time and keep up the good work!